backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.27.0
SQLAlchemy==1.4.51
PyMySQL==1.1.0
cryptography==41.0.7
pydantic>=2.0,<3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openpyxl==3.1.2
python-multipart==0.0.9
python-dotenv==1.0.0
